+++ template/src/modules/services/autoMatchBaseUrl.js
@@ -1,5 +1,3 @@
-import config from 'config';
-
 /**
  * 根据前缀，自动匹配基础的url
  * 根据项目所需，自己扩展
@@ -9,14 +7,11 @@ import config from 'config';
 export default function autoMatchBaseUrl(prefix) {
   let baseUrl = '';
   switch (prefix) {
-    case config.MZS_PREFIX:
-      baseUrl = window.LOCAL_CONFIG.API_MZS;
-      break;
-    case config.UPLOADING_PREFIX:
+    case window.CT.UPLOADING_PREFIX:
       prefix = '';
       baseUrl = window.LOCAL_CONFIG.API_UPLOAD;
       break;
-    case config.OPERATOR_PREFIX:
+    case window.CT.OPERATOR_PREFIX:
       baseUrl = window.LOCAL_CONFIG.API_HSAIR;
       break;
     default: